server/requirements.txt
asgiref==3.5.2
attrs==22.2.0
black==22.10.0
click==8.1.3
colorama==0.4.6
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
iniconfig==2.0.0
mypy-extensions==0.4.3
packaging==23.0
pathspec==0.10.2
platformdirs==2.5.4
pluggy==1.0.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
PyJWT==2.6.0
pytest==7.2.1
pytest-testdox==3.0.1
python-dotenv==0.21.1
pytz==2022.6
sqlparse==0.4.3
